requirements.txt
python-dotenv==1.0.1
PyYAML==6.0.2
Flask==3.1.2
requests>=2.31.0
types-requests>=2.28.14
kubernetes==34.1.0
